Move prettyprint to utils
